+++ classes/vector3.h
@@ -3,15 +3,15 @@
 
 class Vector3 {
 public:
-    int x;
-    int y;
-    int z;
+    float x;
+    float y;
+    float z;
 
-    Vector3(int x, int y, int z);
+    Vector3(float x, float y, float z);
 
-    int scalar_product_of_vectors_this_and_v2(Vector3 *v2);
+    float scalar_product_of_vectors_this_and_v2(Vector3 *v2);
 
-    int len_of_vector3();
+    float len_of_vector3();
 
 private:
